betty checks
